refactor: move contents from lib.rs to digests.rs
